[CAZB-4501] Payment history export api call fix (#1035)
