+++ app/api_wrappers/payment_history_api.rb
@@ -78,32 +78,5 @@ class PaymentHistoryApi < PaymentsApi
       log_action('Getting payment history')
       request(:get, "/payments/#{payment_id}")
     end
-
-    ##
-    # Calls +/v1/accounts/:accountId/payment-history-export+ endpoint with +POST+ method and returns
-    # an empty body response.
-    #
-    # ==== Attributes
-    #
-    # * +account_id+ - uuid, id of the account
-    # * +recipient_id+ - uuid, id of the user who is going to receive an email
-    # * +filtered_user_id+ - uuid, id of the user whose payments are going to be exported
-    #
-    # ==== Result
-    #
-    #   Returns an empty body
-    #
-    # ==== Exceptions
-    #
-    # * {404 Exception}[rdoc-ref:BaseApi::Error404Exception] - account not found
-    # * {500 Exception}[rdoc-ref:BaseApi::Error500Exception] - backend API error
-    #
-    def payment_history_export(account_id:, recipient_id:, filtered_user_id: nil)
-      log_action('Initiating payment history download')
-      body = { recipientAccountUserId: recipient_id }
-      body.merge!(filteredPaymentsForAccountsUserId: filtered_user_id) if filtered_user_id
-
-      request(:post, "/accounts/#{account_id}/payment-history-export", body: body.to_json)
-    end
   end
 end
